invoice-sorter/requirements.txt
# Core Dependencies
python-dateutil>=2.8.2
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Google Cloud APIs
google-cloud-vision>=3.4.0
google-cloud-storage>=2.10.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
pillow>=10.0.0

# OCR and Text Processing
pytesseract>=0.3.10
regex>=2023.8.8

# Web Framework (optional - for dashboard)
flask>=3.0.0
flask-cors>=4.0.0

# LLM Integration (optional)
openai>=1.3.0
anthropic>=0.5.0

# Database (optional)
google-cloud-firestore>=2.13.0

# Utilities
requests>=2.31.0
tenacity>=8.2.3

# Development Dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Logging
structlog>=23.2.0
